transcode: pin wire contracts for all transcoding directions
Replace the loose hand-written wire structs with pinned contract types in
self-contained wire packages (internal/transcode/wire/{object,
openairesponses,openaichat,anthropicmessages}) governed by the new
authoritative contracts.lock.json registry (openai-go v1.12.0 for
openai_responses and openai_chat_completions; anthropic-api 2023-06-01 for
anthropic_messages). pins.md's pinned-revisions table is generated from the
lock (go generate, self-verifying) and covered by a drift test.

The six former wire files (anthropic.go, chat.go, responses_envelope.go,
responses_input.go, responses_output.go, events.go) become re-export shims.
Wire types are distinct per role (create request, response object, each SSE
event, chat request/response/chunk, messages request/response/events) so
request optionality never leaks into response output.

Presence-aware strict decoding (wire.Field[T], wire.Decode) rejects all six
malformed-JSON categories as typed wire.DecodeError: duplicate keys, unknown
fields, missing required fields, illegal nulls, contradictory union arms,
trailing values. Transcode boundaries translate wire.UnsupportedTypeError
into UnsupportedFeatureError before the upstream-wire guard, so valid but
unsupported source features never classify as corrupt upstream wire.

The Responses response envelope always emits the 14 pinned required fields
with explicit null/empty values, and created_at is float64 end-to-end so
fractional timestamps survive decode, stream identity pinning, and
rendering. Responses function tools require an explicit strict
(Field[bool]): strict-less client tools are rejected client-dialect before
any upstream request, and Messages-to-Responses tools emit explicit
strict:false only under the new tool_schema_strictness loss permission.
Every Chat assistant tool call carries type=function in raw wire, and the
rendered create request's stream field is omitted when non-streaming
(omitzero + Field.IsZero) and true only when streaming.

Raw-key golden tests prove the exact emitted key set and nullability per
golden object (never self-oracle), the six decode categories are asserted
typed at every boundary, and the 20 named review-08 regressions stay green
with justified fixture migrations.

+++ internal/transcode/handler.go
@@ -1432,9 +1432,11 @@ func isContextCancellationError(err error) bool {
 	return errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded)
 }
 
-// nowUnix returns the current unix time in seconds.
-func nowUnix() int64 {
-	return time.Now().Unix()
+// nowUnix returns the current unix time in seconds as a float64: the
+// Responses contract's created_at is float64, so the generated stream
+// identity timestamps share the type end-to-end (review-z commit 1).
+func nowUnix() float64 {
+	return float64(time.Now().Unix())
 }
 
 // derefInt returns the value of v, or 0 when v is nil.
